requirements.txt
fastapi
uvicorn
pandas
networkx
python-dotenv
google-generativeai
